add target mask

+++ mask/datav3.py
@@ -35,9 +35,9 @@ def encode(landmarks, phrase, table, max_target_length, start_token="S", end_tok
 def linear_decay(max_value, current_step, total_steps, warmup_ratio=0.1):
     warmup_steps = int(total_steps * warmup_ratio)
     if current_step <= warmup_steps:
-        return current_step * max_value / warmup_steps
+        return tf.cast(current_step, dtype=tf.float32) * max_value / warmup_steps
     else:
-        return max_value - max_value * (current_step - warmup_steps) / (total_steps - warmup_steps)
+        return max_value - max_value * tf.cast((current_step - warmup_steps), dtype=tf.float32) / (total_steps - warmup_steps)
 
 def apply_mask(sample, mask_prob, mask_token_id, random_token_prob, current_step, total_steps):
     mask_prob = tf.cast(linear_decay(mask_prob, current_step, total_steps), dtype=tf.float32)
